feat:iOS性能详情接口

+++ weqa-client-platform-service/src/main/java/com/netease/vcloud/qa/service/perf/PerfTaskStatisticBuild.java
@@ -8,9 +8,9 @@ import com.netease.vcloud.qa.service.perf.view.PerfTaskStatisticVO;
  */
 public class PerfTaskStatisticBuild {
 
-    private Number max = Double.MIN_VALUE;
+    private Number max = Integer.MIN_VALUE;
 
-    private Number min = Double.MAX_VALUE;
+    private Number min = Integer.MAX_VALUE;
 
     private Number total = 0;
